requirements_min.txt
Flask==3.0.3
tinydb==4.8.2
PyJWT==2.9.0